+++ apps/client/src/routes/__root.tsx
@@ -1,7 +1,7 @@
 import { createRootRouteWithContext, Outlet } from '@tanstack/react-router';
 import { TanStackRouterDevtools } from '@tanstack/router-devtools';
 import { QueryClient } from '@tanstack/react-query';
-import { AuthContext } from '@/contexts/authContext';
+import { AuthContext } from '@/features/auth/authContext.tsx';
 
 interface RouterContext {
   auth: AuthContext;
