+++ apps/client/src/routes/_auth.$project.board.tsx
@@ -1,5 +1,5 @@
 import { createFileRoute } from '@tanstack/react-router';
-import Kanban from '@/components/Kanban.tsx';
+import Kanban from '@/shared/components/Kanban.tsx';
 
 export const Route = createFileRoute('/_auth/$project/board')({
   component: Kanban,
