+++ apps/client/src/routes/login.tsx
@@ -1,6 +1,6 @@
 import { createFileRoute, redirect } from '@tanstack/react-router';
 
-import Login from '@/components/Login.tsx';
+import Login from '@/features/auth/Login.tsx';
 
 export const Route = createFileRoute('/login')({
   beforeLoad: ({ context }) => {
