+++ apps/client/src/shared/icon/Harmony.tsx
