+++ apps/client/src/shared/icon/HarmonyWithText.tsx
