+++ apps/client/src/shared/icon/index.ts
@@ -1,0 +1,5 @@
+import Github from '@/shared/icon/GitHub.tsx';
+import HarmonyWithText from '@/shared/icon/HarmonyWithText.tsx';
+import Harmony from '@/shared/icon/Harmony.tsx';
+
+export { HarmonyWithText, Harmony, Github };
