+++ apps/client/src/shared/lib/utils.ts
