+++ apps/client/src/shared/ui/dropdown-menu.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import * as DropdownMenuPrimitive from '@radix-ui/react-dropdown-menu';
-import { cn } from '@/lib/utils';
+import { cn } from '@/shared/lib/utils.ts';
 
 const DropdownMenu = DropdownMenuPrimitive.Root;
